pyproject.toml
[project]
name = "bookstore"
version = "0.1.0"
description = ""
authors = [
    {name = "Anderson Costa",email = "costaandersom23@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pytest (>=8.3.5,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "pytest-django (>=4.11.1,<5.0.0)",
    "faker (>=37.3.0,<38.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
